# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Autobody Shop API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vehicles:
    get:
      operationId: listVehicles
      tags: [vehicles]
      summary: List all vehicles
      parameters:
        - name: clientId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of vehicles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vehicle"
    post:
      operationId: createVehicle
      tags: [vehicles]
      summary: Create a vehicle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VehicleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
  /vehicles/{id}:
    get:
      operationId: getVehicle
      tags: [vehicles]
      summary: Get a vehicle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vehicle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
        "404":
          description: Not found
    patch:
      operationId: updateVehicle
      tags: [vehicles]
      summary: Update a vehicle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VehicleUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
    delete:
      operationId: deleteVehicle
      tags: [vehicles]
      summary: Delete a vehicle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /vehicles/{id}/photos:
    get:
      operationId: listVehiclePhotos
      tags: [vehicles]
      summary: List photos for a vehicle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of photos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VehiclePhoto"
    post:
      operationId: addVehiclePhoto
      tags: [vehicles]
      summary: Add a photo to a vehicle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VehiclePhotoInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VehiclePhoto"

  # --- Check-ins ---
...
  /payments:
    get:
      operationId: listPayments
      tags: [payments]
      summary: List all payments
      parameters:
        - name: clientId
          in: query
          schema:
            type: integer
        - name: checkinId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Payments list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"
    post:
      operationId: createPayment
      tags: [payments]
      summary: Create a payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments/{id}:
    get:
      operationId: getPayment
      tags: [payments]
      summary: Get a payment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
        "404":
          description: Not found
    patch:
      operationId: updatePayment
      tags: [payments]
      summary: Update a payment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"

  # --- Dashboard ---
...
components:
  schemas:
...
    Vehicle:
      type: object
      required: [id, clientId, make, model, year, licensePlate, createdAt]
      properties:
        id:
          type: integer
        clientId:
          type: integer
        make:
          type: string
        model:
          type: string
        year:
          type: integer
        licensePlate:
          type: string
        vin:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
        mileage:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    VehicleInput:
      type: object
      required: [clientId, make, model, year, licensePlate]
      properties:
        clientId:
          type: integer
        make:
          type: string
        model:
          type: string
        year:
          type: integer
        licensePlate:
          type: string
        vin:
          type: string
        color:
          type: string
        mileage:
          type: integer
        notes:
          type: string
    VehicleUpdate:
      type: object
      properties:
        make:
          type: string
        model:
          type: string
        year:
          type: integer
        licensePlate:
          type: string
        vin:
          type: string
        color:
          type: string
        mileage:
          type: integer
        notes:
          type: string
    VehiclePhoto:
      type: object
      required: [id, vehicleId, url, caption, createdAt]
      properties:
        id:
          type: integer
        vehicleId:
          type: integer
        url:
          type: string
        caption:
          type: ["string", "null"]
        createdAt:
          type: string
    VehiclePhotoInput:
      type: object
      required: [url]
      properties:
        url:
          type: string
        caption:
          type: string
...
    Payment:
      type: object
      required: [id, clientId, checkinId, amount, status, method, createdAt]
      properties:
        id:
          type: integer
        clientId:
          type: integer
        checkinId:
          type: integer
        amount:
          type: number
        status:
          type: string
          enum: [pending, paid, partial, refunded]
        method:
          type: ["string", "null"]
          enum: [cash, card, check, financing, null]
        notes:
          type: ["string", "null"]
        paidAt:
          type: ["string", "null"]
        createdAt:
          type: string
    PaymentInput:
      type: object
      required: [clientId, checkinId, amount, status]
      properties:
        clientId:
          type: integer
        checkinId:
          type: integer
        amount:
          type: number
        status:
          type: string
          enum: [pending, paid, partial, refunded]
        method:
          type: string
          enum: [cash, card, check, financing]
        notes:
          type: string
        paidAt:
          type: string
    PaymentUpdate:
      type: object
      properties:
        amount:
          type: number
        status:
          type: string
          enum: [pending, paid, partial, refunded]
        method:
          type: string
          enum: [cash, card, check, financing]
        notes:
          type: string
        paidAt:
          type: string
...