# lib/api-zod/src/generated/types/vehicle.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Autobody Shop API
 * OpenAPI spec version: 0.1.0
 */

export interface Vehicle {
  id: number;
  clientId: number;
  make: string;
  model: string;
  year: number;
  licensePlate: string;
  /** @nullable */
  vin?: string | null;
  /** @nullable */
  color?: string | null;
  /** @nullable */
  mileage?: number | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}