# lib/api-zod/src/generated/types/vehicleUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Autobody Shop API
 * OpenAPI spec version: 0.1.0
 */

export interface VehicleUpdate {
  make?: string;
  model?: string;
  year?: number;
  licensePlate?: string;
  vin?: string;
  color?: string;
  mileage?: number;
  notes?: string;
}